Improve code coverage and fix potential bugs
CommandCanExecuteSourceAttribute was renamend to CanExecuteSourceAttribute and is now only allowed on ICommand.CanExecute() in classes implementing IRaiseCanExecuteChanged.

PropertySourceCollectionAttribute and CommandCanExecuteSourceCollectionAttribute were removed with no replacement as part of a refactoring to reduce code complexity and potential of bugs.

AsyncCommand, AsyncRelayCommand, Command and RelayCommand were removed.

The ViewModel.Children property was removed, consequential ChildViewModelAttribute and ChildViewModelCollectionAttribute were also removed.

In TreeViewModel, if children are present, the TreeChildren property should be overriden for IsChecked to work properly.

Refactor most classes

Adjust namespace of Smaragd.ViewModels.Helpers to Smaragd.Helpers