Fix warnings regarding String.IsNullOrEmpty and nullability
Since the NotNullWhenFalseAttribute doesn't exist when targeting .NET Standard 2.0, the use of String.IsNullOrEmpty produces false warnings since the compiler doesn't know the string was checked for null.
Targeting .NET Standard 2.1 or above would fix this, but that would imply loosing support for .NET Framework. This fixes the warnings by using the ! operator.